backend/requirements.full.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
asteval==1.0.8
astunparse==1.6.3
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.12.2
blinker==1.9.0
boto3==1.42.86
botocore==1.42.86
bw-migrations==0.2
bw2calc==2.4.0
bw2data==4.6
bw2io==0.9.14
bw2parameters==1.1.0
bw_processing==1.0
cached-property==2.0.1
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
chardet==7.4.1
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
constantly==23.10.4
constructive_geometries==1.0
country_converter==1.3.2
cryptography==46.0.7
daphne==4.2.1
dataflows-tabulator==1.54.3
datapackage==1.15.4
deepdiff==7.0.1
Deprecated==1.3.1
Django==6.0.4
django-cors-headers==4.9.0
django-unfold==0.89.0
djangorestframework==3.17.1
docopt==0.6.2
et_xmlfile==2.0.0
flexcache==0.3
flexparser==0.4
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.2
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
ijson==3.5.0
Incremental==24.11.0
iniconfig==2.3.0
isodate==0.7.2
jmespath==1.1.0
jsonlines==4.0.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
linear-tsv==1.1.0
llvmlite==0.47.0
loguru==0.7.3
lxml==6.0.2
matrix_utils==0.7.1
morefs==0.2.2
mrio-common-metadata==0.2.1
msgpack==1.1.2
numba==0.65.0
numpy==1.26.4
oauthlib==3.3.1
openpyxl==3.1.5
ordered-set==4.1.0
packaging==26.0
pandas==3.0.2
peewee==4.0.4
pillow==12.2.0
Pint==0.25.3
platformdirs==4.9.6
pluggy==1.6.0
prettytable==3.17.0
proto-plus==1.27.2
protobuf==5.29.6
psycopg2==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyecospold==4.0.1
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
pyparsing==3.3.2
PyPDF2==3.0.1
PyPrind==2.11.3
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
pyxlsb==1.0.10
PyYAML==6.0.3
randonneur==0.7.2
randonneur_data==0.7.2
RapidFuzz==3.14.5
rdflib==7.6.0
redis==7.4.0
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
rfc3986==2.0.0
rpds-py==0.30.0
sentry-sdk==2.46.0
s3transfer==0.16.0
scipy==1.17.1
service-identity==24.2.0
six==1.17.0
snowflake-id==1.0.2
soupsieve==2.8.3
SPARQLWrapper==2.0.0
sparse==0.18.0
SQLAlchemy==2.0.49
sqlparse==0.5.5
stats_arrays==1.0
structlog==25.5.0
tableschema==1.21.0
toolz==1.1.0
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
ujson==5.12.0
unicodecsv==0.14.1
uritemplate==4.2.0
urllib3==2.6.3
voluptuous==0.16.0
wcwidth==0.6.0
wheel==0.46.3
win32_setctime==1.2.0
workos==5.46.0
wrapt==2.1.2
wurst==0.5.3
xarray==2026.2.0
xlrd==2.0.2
xlsxwriter==3.2.9
zope.interface==8.2
